decompose network filters
